+++ packages/editor/src/extensions/rich-text/RichTextPlugin.tsx
@@ -1,13 +1,6 @@
 import { TabIndentationPlugin } from '@lexical/react/LexicalTabIndentationPlugin';
 
-import DraggableBlockPlugin from './DraggableBlockPlugin';
-
-const RichTextPlugin: React.FC = () => (
-  <>
-    <TabIndentationPlugin />
-    <DraggableBlockPlugin />
-  </>
-);
+const RichTextPlugin: React.FC = () => <TabIndentationPlugin />;
 
 RichTextPlugin.displayName = 'extensionRichText.Plugin';