+++ packages/editor/src/index.dev.tsx
@@ -7,6 +7,6 @@ const container = document.getElementById('root')!;
 const root = ReactDOM.createRoot(container);
 root.render(
   <React.StrictMode>
-    <Editor minHeight="calc(100vh - 65px)" />
+    <Editor minHeight="calc(100vh - 57px)" autofocus />
   </React.StrictMode>
 );